+++ src/pages/Auth/Dashboard/AddTransaction/index.tsx
@@ -2,9 +2,10 @@ import { useState } from 'react';
 import { Close } from '@mui/icons-material';
 import { Box } from '@mui/system';
 import * as S from './styles';
-import { Form } from './components/Form';
-import { Success } from './components/Success';
+import { Form } from './Form';
+import { Success } from './Success';
 import { useModal } from '@/context/modal';
+import { Modal } from '@mui/material';
 
 export const AddTransaction = () => {
   const { toggleModal } = useModal();
@@ -24,45 +25,47 @@ export const AddTransaction = () => {
   };
 
   return (
-    <Box
-      component="div"
-      sx={{
-        top: '10vh',
-        maxWidth: '478px',
-        ml: 'auto',
-        mr: 'auto',
-        backgroundColor: 'transparent',
-        position: 'absolute',
-        left: '50%',
-        transform: 'translate(-50%, 0)',
-        width: '100%',
-        height: '100vh',
-      }}
-    >
+    <Modal open sx={{ height: '100vh' }}>
       <Box
         component="div"
-        bgcolor="#2B2E35"
         sx={{
-          display: 'block',
-          px: 2.5,
-          mb: 2,
-          borderRadius: 6,
-          py: 2,
-          paddingBottom: 3,
-          height: '90%',
-          overflowX: 'scroll',
+          top: '10vh',
+          maxWidth: '478px',
+          ml: 'auto',
+          mr: 'auto',
+          backgroundColor: 'transparent',
+          position: 'absolute',
+          left: '50%',
+          transform: 'translate(-50%, 0)',
+          width: '100%',
+          height: '100vh',
         }}
       >
-        <S.Container>
-          <S.Header>
-            <button type="button" onClick={() => handleCloseModal()}>
-              <Close />
-            </button>
-            <h1>Adicionar Transação</h1>
-          </S.Header>
-          {components[step]}
-        </S.Container>
+        <Box
+          component="div"
+          bgcolor="#2B2E35"
+          sx={{
+            display: 'block',
+            px: 2.5,
+            mb: 2,
+            borderRadius: 6,
+            py: 2,
+            paddingBottom: 3,
+            height: '90%',
+            overflowX: 'scroll',
+          }}
+        >
+          <S.Container>
+            <S.Header>
+              <button type="button" onClick={() => handleCloseModal()}>
+                <Close />
+              </button>
+              <h1>Adicionar Transação</h1>
+            </S.Header>
+            {components[step]}
+          </S.Container>
+        </Box>
       </Box>
-    </Box>
+    </Modal>
   );
 };
